finxing uloadign

+++ resources/views/layouts/app.blade.php
@@ -33,6 +33,7 @@
     <!-- INTERNAL Switcher css -->
     <link href="{{ asset('admin') }}/switcher/css/switcher.css" rel="stylesheet">
     <link href="{{ asset('admin') }}/switcher/demo.css" rel="stylesheet">
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/jquery-toast-plugin/1.3.2/jquery.toast.min.css" rel="stylesheet">
 
 </head>
 
@@ -108,6 +109,7 @@
 
         <!-- Switcher js -->
         <script src="{{ asset('admin') }}/switcher/js/switcher.js"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-toast-plugin/1.3.2/jquery.toast.min.js"></script>
         <script>
             class DataTable {
                 constructor(tableId, url, columns) {
@@ -135,6 +137,7 @@
                         success: () => {
                             $(`#${this.tableId}`).DataTable().ajax.reload();
                             $('#modal-form').modal('hide');
+                            toast('Data berhasil di tambahkan','success')
                         }
                     });
                 }
@@ -149,7 +152,6 @@
                             // Menampilkan data pada modal atau form
                             $('#modal').modal('show');
                             $('#name').val(data.name);
-                            // ...
                         }
                     });
                 }
@@ -164,22 +166,61 @@
                         success: () => {
                             $(`#${this.tableId}`).DataTable().ajax.reload();
                             $('#modal-form').modal('hide');
+                            toast('Data berhasil di perbarui','success')
                         }
                     });
                 }
 
-                delete(id) {
+                delete(url) {
                     // Mengirim data ke backend untuk melakukan delete
                     // Setelah berhasil, melakukan refresh pada tabel
-                    $.ajax({
-                        url: '',
-                        method: 'DELETE',
-                        data: { id: id },
-                        success: () => {
-                            $(`#${this.tableId}`).DataTable().ajax.reload();
+                    Swal.fire({
+                        title: 'Are you sure?',
+                        text: "You won't be able to revert this!",
+                        icon: 'warning',
+                        showCancelButton: true,
+                        confirmButtonColor: '#3085d6',
+                        cancelButtonColor: '#d33',
+                        confirmButtonText: 'Yes, delete it!'
+                        }).then((result) => {
+                        if (result.isConfirmed) {
+                            $.ajax({
+                                url: url,
+                                method: 'DELETE',
+                                success: () => {
+                                    $(`#${this.tableId}`).DataTable().ajax.reload();
+                                    toast('Data berhasil di hapus','success')
+                                }
+                            });
+                            Swal.fire(
+                                'Deleted!',
+                                'Your file has been deleted.',
+                                'success'
+                            )
                         }
-                    });
+                    })
+                    
                 }
+            }
+            function toast(text,icon){
+                $.toast({
+                    text: text, // Text that is to be shown in the toast
+                    heading: 'Note', // Optional heading to be shown on the toast
+                    icon: icon, // Type of toast icon
+                    showHideTransition: 'fade', // fade, slide or plain
+                    allowToastClose: true, // Boolean value true or false
+                    hideAfter: 3000, // false to make it sticky or number representing the miliseconds as time after which toast needs to be hidden
+                    stack: 5, // false if there should be only one toast at a time or a number representing the maximum number of toasts to be shown at a time
+                    position: 'bottom-left', // bottom-left or bottom-right or bottom-center or top-left or top-right or top-center or mid-center or an object representing the left, right, top, bottom values
+                    
+                    textAlign: 'left',  // Text alignment i.e. left, right or center
+                    loader: true,  // Whether to show loader or not. True by default
+                    loaderBg: '#9EC600',  // Background color of the toast loader
+                    beforeShow: function () {}, // will be triggered before the toast is shown
+                    afterShown: function () {}, // will be triggered after the toat has been shown
+                    beforeHide: function () {}, // will be triggered before the toast gets hidden
+                    afterHidden: function () {}  // will be triggered after the toast has been hidden
+                });
             }
         </script>
         @yield('js')
